HSEARCH-646 Deleting test. Test outcome were dependent upon the environment.
A real test would need to test scenarios. Once a lock timeout in the mass indexer and once a lock timeout during "normal" indexing. In both cases one has to check that the index is not corrupted afterwards. One way to tests this could be to utilize Byteman, but the setup is quite tricky. For now we delete this test.