HSEARCH-780 Fix for @IndexedEmbedded PersistentSet dirty analysis in a super class
Wrong collectionRole is used for collection exclusion analysis in
FullTextIndexEventListener.processCollectionEvent > AbstractDocumentBuilder.isCollectionRoleExcluded check.

A PersistentSet of an abstract superclass has it's collection role bound to the superclass, but AbstractDocumentBuilder's
indexedEmbeddedCollectionRoles contains roles that are bound to a specific subclass. Thus when the contents of a collection
in superclass are changed, Hibernate Search fails to re-index the change as the collection gets excluded.