HSEARCH-921 Removing failing 'fullFileName.replace( '/', File.separatorChar )' call. Since we are loading from the classpath this is not needed
Also moving the file and resource handling into a static helper class (makes it reusable and testable). Access to classloader is also unified in one spot. This might be useful in case we need to revisit how we use classloaders.

Handling the case where a unknown resource is specified with an exception