HSEARCH-2301 Use the most specific type to build the criteria
When building the criteria to get the entities, we used to always use
the root entity type of the hierarchy even if we only want to get
entities from one subtype.

This is especially counter productive when we are in the case of a join
inheritance with many subtypes: we then build a query with a left join
for each subtype instead of a query only targeting the most specific
subtype possible.