HSEARCH-2260 Fixed an issue with index creation timing out
When the index name contained some special character that should be url-encoded (such as '$'), it was not url-encoded.
As a result, we waited for the creation of an index that was never created in the first place, so we were reaching the timeout and considered the index creation had failed, whereas it had not.