HSEARCH-2352 Fixed an issue with Index.NO not being taken into account (ES integration)
It was ignored on booleans, numerics and dates because of a control that was too strict.

Also, because Index.NO was being ignored, a bug on non-indexed fields had not been spotted before: sort on non-indexed fields did not work. This was solved by adding doc_values to non-indexed, sortable fields.