HSEARCH-2396 Support for creating nested fields from custom field bridges with Elasticsearch
In order to do that, we have to make sure to properly analyze the end of
the field names (the part that is not an @IndexedEmbedded prefix), and
create nodes for path components in this part of the path.

The implied changes to the implementation are not trivial, because the
algorithms in JsonTreeBuilder that previously always created empty
JsonObjects now must also be able to add a user-supplied JsonPrimitive.
The introduction of JsonAccessor helps to avoid code duplication and
to keep the code simple.