HSEARCH-2396 Enable indexing the ID as part of the document with ES
The ID was previously only provided to ES in the URL, resulting in
searches on the ID field with the Hibernate Search field name
(instead of the ES-specific '_id' keyword) failing miserably.

Also, re-enabled some tests that used not to work on ES due to the lack
of support for ID searches.