HSEARCH-2393 Fixed ProgrammaticMappingTest.testFullTextFilterDefAtMappingLevel
We don't need advanced filters, we just test that the filter we
configured is actually used. Which is why replacing the actual filter
implementation seems okay.

Also, we don't need to test analysis on the "owner" field, so we better
use a simple keyword and avoid conflicts between the way the Lucene
backend and the Elasticsearch backend do the analysis.