HSEARCH-2360 Use source filtering in Elasticsearch queries
To achieve that, we need to inspect the metadata for each projected
field before querying. Since the code was very similar to the one
executed when extracting query results, I chose to build a "projection
metamodel" during the preliminary inspection projected field metadata.
That removes the need of inspecting metadata when extracting query hits.