HSEARCH-2480 Restored pre-elasticsearch parameters for MutableFactoryTest.testMultiThreadedAddClasses
The number of tasks and threads had been lowered because of timeouts
that I thought were due to Elasticsearch being too slow for the test to
work, but it appears that it was related to the bug we just fixed.

Thus we can raise the number of tasks and threads again.