HSEARCH-2448 Make metadata order as stable as possible
It helps when debugging and writing tests.

By using LinkedHashSet/LinkedHashMap instead of HashSet/HashMap, we have
at least the guarantee that the metadata order will match the order in
which they were processed by the annotation processor. It does not mean
the order is fully predictable, since we process class fields/methods in
the same order they are returned by the JVM (which is undefined), but at
least we can rely on some properties of the annotation processor, such
as the fact that methods are handled before fields.