HSEARCH-2415 Make it possible for any field bridge to define null encoding and indexNullAs parsing
Previously, we used type introspection on the field bridges to determine
whether a field bridge was using numeric encoding on its main (default)
field and to parse indexNullAs accordingly. The downside of this method
is that it prevents extensions from additional modules such as the
Elasticsearch integration.

This commit introduces the EncodingBridge interface for bridges with
specific data encoding on the main field, making it possible for those
to define what their numeric encoding is (if any) and how to parse their
indexNullAs parameter.