HSEARCH-2554 Only wrap bridges in container bridges when the new @IndexedContainer annotation is used
For fields on container properties (array, iterable, map) that were only
annotated with @Field, the behavior is unchanged: field bridges will be
passed the container value directly, not the element value.

This allows us to avoid breaking custom container bridges
(HSEARCH-2508).

For fields that were annotated with both @Field and @IndexedEmbedded,
the behavior will change: field bridges will now be passed the container
value directly, not the element value as they used to. To revert to the
previous behavior, the @IndexedContainer annotation should be added,
using the forField attribute as necessary. The indexNullAs may be
defined on the @IndexedContainer annotation directly.

This allows us to provide a way for users of the Elasticsearch
integration to index containers easily and to use indexNullAs on
both the containers and their elements (HSEARCH-2419).

Note that another side-effect is that we now use proper encoding when
indexing null tokens for containers. I.e. if the elements are numeric,
the null token for the container is expected to be numeric too.

While this is a breaking change, the feature that we break (container indexing
using element field bridges) was not documented and arguably already broken,
since we required to use @IndexedEmbedded to enable it, which in turn
enables another feature: the embedding of fields carried by the
elements in the root document.