HSEARCH-2527 Make Scoped*AnalyzerReferences drop the reference to their builder ASAP
This allows to have one less map in memory at runtime for each scoped
analyzer reference. Maybe not worth it, but at least it's cleaner...

To achieve that, we have to build copied scoped analyzer without the
original map of analyzer references by scope, which means the copy
builder becomes quite different from the initial builder, and that we
must retrieve the global analyzer and analyzers by scope from the
scoped analyzer.