HSEARCH-2571 Remove duplicate imports causing Karaf to reject the OSGi bundle
It seems the bug causing the maven-bundle-plugin to forget to include
some imports changed a bit, and it now only forgets about
org.hibernate.search.query.dsl.sort for some reason.

Thus we can remove the explicit import for other -engine packages, and
in fact we must, because otherwise we get duplicate imports, and Karaf
doesn't seem to like that.