HSEARCH-2418 Make ElasticsearchAnalyzer responsible for registering analyzer definitions in the settings
This will allow us to have some analyzers hold a reference to
an ES-specific analyzer definition, so that we can have analyzers
definition coming from another source than just @AnalyzerDef
annotations.

There are a few reasons to do that, the most important being that the
@AnalyzerDef annotation does not really fit the ES model, where one
tokenizer definition (for instance) can be referenced from multiple
analyzer definitions.

One side-effect of this change is the translation from Lucene class
providing from @Analyzer annotations to the Elasticsearch name must now
be done when initializing references, instead of when building the ES
schema. This is inconsistent with the translation of @AnalyzerDef
annotations, which occurs when building the ES schema, but we'll fix
this inconsistency in the next commits.