HSEARCH-2453 Add dedicated configuration options for Elasticsearch authentication
Authentication was already possible, but a bit painful, since you had to
include the username and password in the host URLs. Which meant
repeating it over and over again when using multiple URLs, and also
meant you had to URL-encode the username and password yourself.

Tests are not included, because it's a pain to automate the setup of
authentication on the ES instance used for tests. Also, I'm quite sure
we can't do it with an embedded version of ES, and I don't even know if
the licensing terms would allow it (the Shield plugin is subject to a
commercial license).

To test manually, see:
https://www.elastic.co/blog/getting-started-with-elasticsearch-ssl-native-authentication