Remove some dead code
If we reach the "if ( isNumeric(...)) " in "getType",
it means we don't have any bridge-defined field for this
field name (see the code at the top of "getType").
Thus "FieldHelper.getNumericEncoding(...)" is useless, we can simply
use "DocumentFieldMetadata.getNumericEncoding()".

Also, "DocumentFieldMetadata.isNumeric" already is true if the field
bridge is a NumericFieldBridge or wraps one.
Thus "FieldHelper.isNumeric(...)" is useless, we can simply use
"DocumentFieldMetadata.isNumeric()".