HSEARCH-2541 Don't use Session.load to load objects, use Session.get() instead
Session.get() *actually* loads the object, while Session.load() only
creates a proxy. That's terrible naming, but it probably is legacy API.

In our case, we ultimately want fully initialized objects, so
Session.get() is more appropriate. On top of that, Hibernate ORM seems
to have issues with initializing proxies on entities with public fields,
so that's all the more reason to use Session.get() and bypass proxying
altogether.