HSEARCH-2581 Use at least as many async ES connection threads as the maximum number of connections
The RestClient uses the async HTTP client even for synchronous calls,
so the number of threads in the async HTTP client basically defines
the maximum number of concurrent Elasticsearch calls.

Thus we'd better make it at least as high as the maximum number of
connections, or we'll have unused connections.