HSEARCH-1316 Allow integrators to defer the initialization of the search factory using a provided EnvironmentSynchronizer
This is necessary to integrate with CDI on WildFly in particular.

When WildFly starts Hibernate ORM, the CDI context has already
been bootstrapped, but it hasn't been initialized, so one cannot request
beans yet. The CDI context will only get fully initialized *after* Hibernate
ORM has finished starting.

When we want to provide a CDI implementation of BeanResolver, this
creates a cyclic dependency: Hibernate Search needs the BeanResolver to
be available when it starts, but the BeanResolver needs CDI to be
initialized, and CDI won't be initialized before Hibernate ORM has
started, which won't happen before Hibernate Search has started.

The EnvironmentSynchronizer provides a way to break this cyclic
dependency, by defering the initialization of Hibernate Search to when
CDI is fully initialized. If Search is accessed (either to index
entities or to execute search queries) before that, the calls will just
block.