HSEARCH-2382 Merge the JPA and Hibernate-native FullTextQuery implementations
This doesn't change the interfaces, but may change the exception
handling.

Why? Hibernate ORM and JPA haven't always had consistent behaviors, but
it used not to matter because there were different APIs that could be
implemented differently anyway. Now that the two implementations are
merged, there has some choice to be made, and unfortunately this hasn't
been done in a fully consistent way in Hibernate ORM.

Here are some quirks in Hibernate ORM:

 * Some JPA methods behave according to the JPA spec, unless Hibernate
was booted using its native booting mode
(getProducer().getFactory().getSessionFactoryOptions().isJpaBootstrap()
returns false), in which case they may throw Hibernate exceptions.
This is the case for getSingleResult() in particular.
 * Some Hibernate native methods behave according to the JPA spec,
regardless of the booting mode. For instance list() will always convert
Hibernate exceptions to JPA ones.

With that in mind, it's not easy to find a consistent behavior for our
own methods.

Here are some rules I followed:

 * For methods that were already implemented in Hibernate ORM, and whose
implementation was suitable for Hibernate Search, I just delegated to the
Hibernate ORM implementation and didn't override anything.
 * For methods that we have to re-implement in Hibernate Search (e.g.
list()), I replicated the exact same exception handling and delegated to
Hibernate ORM when possible (e.g. for exception conversion), even if we
used to have our own exception conversion code. That's because I suspect
this code had been copied over from an older Hibernate ORM
implementation.
 * For Hibernate Search-specific methods that used to have different
behaviors in the native and JPA implementations, I tried to favor the
JPA behavior regardless of the booting mode. For instance getResultSize()
will throw JPA exceptions, not native ones. If we must break stuff, we
may as well do it in a standard way.