HSEARCH-1886 Introduce IndexManagerGroupHolder, responsible for shards of a single index
We now have a component modeling a "group of indexes", i.e. a group
of shards for a same "logical" index, which will be useful to start
eagerly the JGroups listener on master, even if there are no index
managers yet.

Also, this reduces the complexity of dynamic sharding, and makes it
closer to non-dynamic sharding.