HSEARCH-1886 Remove the reliance on the nodeSelector in JGroupsMasterMessageListener
Rationale:

 1. When using dynamic sharding, we may receive messages before the
target index manager has even been created, because the first entity to
be added to the index was created by a slave. In that case, there isn't
any node selector for this index (yet), and the message ends up being
swallowed.
 2. This check was only necessary to handle cases where the master just
became a slave, and a message was received by the old master just during
the switching. But in this case, the underlying queue processor is
supposed to have switched mode too, and *will* ultimately re-send the
message to the new master... So the check is simply redundant.