HSEARCH-2723 Refactor the documentation to make built-in bridges more visible
We used to have a partial duplicate of the built-in bridges
documentation in the part about @Field, making it easy to miss part of
the documentation (including stuff about dates).

I removed the duplicate and merged it into the built-in bridges
documentation, then linked from the @Field documentation to the built-in
bridges documentation. This should make it less likely that someone
misses parts of the built-in bridges documentation.