HSEARCH-2723 Remove the warning about resolution on @DocumentId
We do not check this at runtime, so it is not "impossible" to have a
document ID with a date bridge whose resolution is lower than the
milisecond. It could even work if all the values have a zero time
(00:00), for instance.

I'm not saying it makes sense, but it's about as dangerous as putting a
random custom field bridge on the document ID without taking care of
preserving information, so _if_ we warn about this, we should not warn
about this _here_.