HSEARCH-2724 Clarify that "directory-based" indexes are actually just "embedded Lucene" indexes
About the log message: I did not add any recommendation as to how to fix
the issue. That's because you can only get this error two ways

 1. You tried to query an entity which is mapped to a custom (non-lucene,
non-elasticsearch) index manager which does not extend
DirectoryBasedIndexManager.
 2. You tried to call SearchIntegrator.createHSQuery(), which is a
deprecated SPI that _we_ don't use internally, so you had to call
it directly.

In both cases, it's advanced use, and I simply cannot give a one-liner
solution.