HSEARCH-2716 WorkerTestCase running out of physical JDBC connections as MariaDB requires twice as much
The default Id generation strategy for `@GeneratedValue` on MariaDB is the
SequenceStyleGenerator, which will require a second JDBC connection to be opened
in isolation from the primary connection os the Session.

This implies that limiting the connection pool to 15 isn't enough when the tests
open 15 Sessions; we need at least double of that.

I'm adding one extra connection so to ensure we can drop the table and perform
cleanup even in case of critical test failure.