HSEARCH-1316 Protect all state depending on the Integrator consistently
The FullTextIndexEventListener has several more fields which depend on the
initialization of the final ExtendedSearchIntegrator, this patch is making
sure that any such necessary state is accessed consistently and via the
required visibility guards.

There's something dodgy going on with Serializability of this class but that
appears to be a problem we already had.