HSEARCH-2649 Retrieve JobOperator from BatchRuntime
Previously, we retrieved the JobOperator from BatchRuntime only once and
stored this reference as a variable inside the JobFactory. It aimed to
avoid the following exception:

    java.util.ServiceConfigurationError:
        javax.batch.operations.JobOperator:
	    Provider org.jberet.operations.JobOperatorImpl could not be instantiated

This happened in JBeret SE, but it does not happen in JBatch SE. There's
no evidence that the JobOperator can be retrieved once and only once
from the BatchRuntime in Java SE. So it's rather a bug of JBeret SE and
we should avoid using JobFactory, the specific workaround solution in our
codebase.