HSEARCH-2791 Do not rely on Hibernate ORM for ORM-unrelated tests
Using ORM brings more complexity to the tests, and with it a greater
potential for failures unrelated to the feature being tested.

For instance, time types are not supported as well, nor in the same way,
with all databases, so we may as well avoid the hassle.

So when we're testing features that are not related to ORM in any way,
let's avoid it.

The tests modified by this commit all failed with one database or
another, most of the time because of entity names that were too long or
because of time types that were not supported the same way as in H2.