HSEARCH-2481 Do not allow the maven-failsafe-plugin to scan the whole classpath for tests
This avoids in particular a problem where byteman classes were being
loaded before the byteman agent was launched, which resulted in two
versions of byteman classes being present: one loaded by the system
classloader (when failsafe scanned the whole classpath), and one when
the byteman agent was launched and added its JAR to the bootstrap
classloader.

For the record, the class that triggered the loading of byteman classes
was org.hibernate.search.testsupport.BytemanHelper, which references
org.jboss.byteman.rule.Rule in its constructor (among others, but
that's the main problem).