HSEARCH-2761 Ignore conflicts in ES5 delete-by-query
Rationale: in our case, Elasticsearch is not our source of truth, some
other component (traditionally a database) is. When we perform a
delete-by-query, it's only to do an index purge, and rebuild the index
after that.
If there is a write to the index concurrently to the purge, it's
generally no big deal: it mainly means the document will be reindexed
just a moment later.