HSEARCH-2803 Change the default value of maxThreads
Job parameter `maxThreads` is used to specify the maximum number of
threads on which to execute the partitions of 2nd step: the production
of Lucene documents. The default was set to 10 be default; now, the
default value is the number of partitions, which is more aggressive.

Note the batch runtime cannot guarantee the requested number of threads
are available; it will use as many as it can up to the requested
maximum.