HSEARCH-2615 Cache Hibernate ORM metadata (ID order etc.) as much as possible
So that we'll avoid checking out the metadata for every single query.

This cache will be most effective on JSR-352 runtimes that actually share
the job context between partitions, such as JBeret; others such as
JBatch force us to rebuild the context for each partition, in which case
the only benefic will be to only rebuild the metadata once for both
the "greater-or-equal" and "lesser" restrictions.