HSEARCH-2764 Avoid a very rare race condition in async work processing
If, between the end of the processing loop and the call to
processingScheduled.set( false ) at the end of processing, another
thread somehow managed to submit a changeset and call awaitCompletion(),
then this thread ended up not waiting for its changeset, but only for
the previous ones.

This commit fixes the issue by avoiding the use of multiple instances of
CountDownLatch, and instead relying on Phaser so that we can safely
change what waiting threads are waiting for (i.e. we can just say
"oh sorry, you were waiting for the previous runnable, but another one
needs to be ran before I let you go"). Also part of the solution is
systematically checking whether a new processing runnable must be
scheduled before arriving at the phaser.