HSEARCH-2764 Make Elasticsearch stream orchestrators index-specific
The downside is we may not be able to bulk as much as we used to, but
there are a few advantages too:

1. We're finally able to force executing synchronous works in order
(by using one serial orchestrator per index). Note that this may impact
performance negatively, but at least we'll avoid some errors.
2. We can finally disable the 'refresh' in bulk API calls when
'refreshAfterWrite' is disabled for the index. Previously we couldn't,
because this parameter can take a different value for each index
manager.