HSEARCH-2616 JSR-352: Implement checkpoints so as to recover correctly from failures
The only safe strategy (at least for now) is to use update works instead
of add works for the first set of writes of every writer.

This could be optimized if we had some way to check wether we are indeed
restarting or we are performing the very first run.
Then we could avoid using update works on the very first job execution.
Unfortunately, if a failure occurs during the very first set of writes
(before the first checkpoint), the JSR-352 won't give us any clue when
restarting: it will be as if it was the very first execution.

Another option would be to perform a search query on the very first
write, checking whether the elements we're about to write are already in
the index or not, and choosing between update or add works accordingly.
This would at least allow us to only execute one search query for all
the documents in this first set of writes, instead of one query per
document. The query would be more complex though (with junctions), so
I'm not sure it would actually be faster. We could give it a try,
though.