HSEARCH-2854 Avoid 0-length writes in GsonHttpEntity
When content-length is defined, the encoder is a LengthDelimitedEncoder,
and even a 0-length write will trigger an exception when the content
length has been reached.

It turns out the issue is very unlikely to occur right now, because it
requires that the content-length be defined (which currently only happens
if the content length is less than 1024 bytes) *and* that the flow
control push back (which is apparently unlikely to happen before the
first 1024 bytes have been written).
But a few changes are coming, mainly HSEARCH-2849, which will improve
content-length detection and thus make the issue more likely to occur.