HSEARCH-2868 Do not ignore updates on un-initialized collections in FullTextIndexEventListener
Some collections, such as the reverse side of an unordered,
duplicate-allowing association, may not be initialized upon adds, and in
those cases we may still want to trigger reindexing, provided the
collection is used to generate an index field.

This fixes HSEARCH-782 in a (hopefully) more correct way. The original
fix added an "if" skipping reindexing when a collection being updated
wasn't initialized, but this is incorrect, as the collection may still
affect the indexed form of the entity (the collection may not be
initialized but still have pending updates in an internal queue).
The new fix removes this skipping, and simply does not add ContainedIn
collection roles to the list of collection roles used when generating
documents from an entity.