HSEARCH-2911 Use one session/scroll per chunk in JSR-352 mass indexer
The JSR-352 spec clearly states that one (JTA) transaction is created
per chunk, and this transaction is committed just after the chunk.
For that reason, re-using a scroll outside a chunk's boundaries will
lead to errors with some databases (PostgreSQL for example), because the
scroll is cleared on the server side when the transaction is committed.
Strangely, H2 seems to be supporting this just fine, but it's not
something we can rely on.

As a result, I had to adjust the way custom HQL queries are handled. In
that mode, we used to run a single query for the whole mass indexing,
but this is no longer possible since we have to split the processing in
as many transactions as there are chunks.
I ended up storing the index of the last result processed in each chunk,
and using that as an offset for the query for the next chunk. This is
admittedly not a perfect solution, but I can't see a better one for now.