HSEARCH-2609 Add periodic session clearing during chunks in JSR-352 mass indexer
Since we flush to the index at the end of each chunk, and an index flush
requires to hold a global lock on the index writer, it is not a sensible
strategy to just use small chunks.

Introducing periodic session clears during the processing of a chunk
will allow to raise the checkpoint interval without running out of memory,
and thus to flush to the index less often. This gives a huge
performance boost (indexing time divided by two on the ORM mass indexing
performance test with PostgreSQL on my machine).