HSEARCH-2932 Increase timeouts in JSR-352 ITS to avoid failures on slow platforms
The symptoms were that some requests started failing exactly 10 seconds
after starting the job because of non-existing index.

Incidentally, 10s was the job timeout. What happened was the test
stopped waiting for the job to complete, and closed the SessionFactory
brutally, resulting in the indexes being automatically dropped
before the pending works had a chance to execute (see HSEARCH-2934).