HSEARCH-3302 Rename some bridge-related types to be more consistent
The new names are based on the ones chosen during the F2F, but are a bit
more consistent. In particular:

* when binding a bridge (when calling "contribute"), an instance of
IndexSchemaElement allows to manipulate part of the index schema and an
instance of PojoModelElement allows to manipulate part of the Pojo model.
** We no longer mention "bridged" in the type names, because these types do
not offer any feature directly related to bridges, only reflection
and read/write features. They could theoretically be used outside of the
context of a Bridge.
** We do mention, however, that this type is related to the POJO model.
** We use PojoModelElement instead of PojoElementModel to mirror the
structure of the name of the other type, "IndexSchemaElement".
* EngineHandle has been renamed to SearchModel to fall in line with
IndexSchemaElement and PojoModelElement.
* the FunctionBridge now has "toIndexValue" and "fromIndexValue" methods,
which is a bit less awkward than "apply"/"applyReverse".
* similarly, the IdentifierBridge has a "toDocumentIdentifier" and a
"fromDocumentIdentifier" method.
* similarly, the RoutingKeyBridge now has a "toRoutingKey" method.
* as a consequence, some backend-related types were renamed
for consistency. In particular classes dealing with the index schema
using the "IndexSchema" wording and classes dealing with more than just
the schema (index name, settings, ...) use the "IndexModel" wording.