HSEARCH-2955 Do not throw an exception when matched documents do not have any facet value
We used to rely on the metadata inside the Lucene index to determine
whether facets are correctly configured, but this is no longer
necessary: we can just check our own metadata. So now if we encounter a
case where Lucene tells us "there are no facets in this set of
documents!", we can simply infer that the documents do not have any
value for the requested facet field.