HSEARCH-3302 Move backend tests to a mapper-independent TCK
The TCK will allow us to write tests once and run them against every
backend.
Stubbing the mapper allows us to write backend tests once instead of
once per mapper. It also makes tests easier to write.

Ultimately, this will allow us us to keep good coverage in our tests
without thoroughly testing all the combinations mapper-backend.

This is important because the combinations can be numerous: we
envision at least two mappers (Hibernate ORM Pojos and free-form),
one of which has two ways of mapping (annotations and programmatic),
and at least two backends (Lucene and Elasticsearch).
If we were to test backends with each mapper, that would already
require 3 copies of each backend test. For some tests, such as
testing that the generated schema is the one we expect, we would also
need backend specifics in the test, so that's 3*2=6 copies of each test.