HSEARCH-3302 Remove the need for ConcurrentReferenceHashMap for storing pojo workers for in-progress transactions
This class was huge and had no unit test. Besides, Sanne wanted to get
rid of it, so let's try something.

The fact that the map is now stored as a transient property inside the
session shouldn't lead to any *new* problem, since we used to rely on a
map with weak keys, and the transaction identifier inside the session is
transient. So previously, when a Session was serialized, the transaction
identifier was only referenced from the map, and thus the value
(containing all the works we want to process on commit) was lost.