HSEARCH-3302 Remove IndexedTypeIdentifier and TypeOrdering and use TypeModel instead
This mainly allows to make it clearer that indexed types (types mapped
to an index) in the Pojo mapper have stricter requirements than other
types. It still isn't crystal-clear, but we're getting closer...

Also, this would theoretically allow a Pojo mapper implementation to
register mappings that are not tied to a java.lang.Class<?>, but to
a java.lang.reflect.Type, which could be a parameterized version of
a class such as List<String> or MyPojo<Foo>, as long as the mapping
is only used in IndexedEmbeddeds and not as an indexed type (type mapped
directly to an index).