HSEARCH-3302 Move the responsibility of extracting type arguments to the PojoTypeModel
This way, we will be able to implement a better type model in the
future, preserving generics type mapping information as we navigate down
the indexed type's property tree.

We may need to rely on a more advanced library for that, though, so it's
better if the type manipulation is isolated in the type model: we may be
able to provide an alternate type model implementation in a separate
module with additional dependencies.