HSEARCH-3302 Remove the use of casts in PojoElement
We already check that the expected type matches the actual type during
bootstrap, so the caster would only be useful if those checks were
incorrect and the property handles or container value extractors return
objects of a different type than what they advertise. This would mean a
bug in property handles or container value extractors, which could
affect much more than just the PojoElements.

If we really want to protect ourselves against such bugs, we should do
it more extensively. I don't think we want this, though, since the casts
would be useless in most cases, and would only help to debug when they
are useful (failures would occur even without the casts, just not in
obvious places).