HSEARCH-3302 Fix XML file headers (POM files in particular)
The files are apparently excluded from the checkstyle scope, because
checkstyle only considers sources and resources by default.
We could make it consider pom files too, but that would require some
exotic configuration, such as including the whole project in the
"sourceDirectories" setting. So I'd rather not go that way.