HSEARCH-3083 Make Closer.push(ClosingOperator, Object) null-safe and remove the null-unsafe method Closer.push(GenericCloseable)
Calling Closer.push(myCloseable::myMethod) is inherently null-unsafe,
since it will fail when myCloseable is null, and can always be replaced
with Closer.push(MyClass::myMethod, myCloseable), which we just made
null-safe.

This change should make it far less likely that we add a NPE by mistake
just because we forgot to check whether an object to close is null.