HSEARCH-3177 Merge MapperFactory and MetadataProvider into a single, clearer class: MappingInitiator
The idea is to make the entry point for mapping bootstrapping more
straightforward.

Also, this removes the previous possibility one had to use two
"MetadataProviders" (now "MappingInitiator") to contribute to a single
mapping. This was dodgy because, even though the two components could
contribute to the same mapping, each would provide its own mapping
factory, and in the end the engine would only use one of them. So if the
mapper factories were different, we could end up with unpredictable
behavior. Now it's safer: whenever you try to contribute to the same
mapping through two different initiators, an exception will be thrown.