HSEARCH-3267 Expose the data source name through WildFlyConfigurationExtension in WF tests
We may decide in the future to use a different datasource in WF tests
depending on build options; I almost had to do it to solve this issue.
So let's be ready for that by moving the DATA_SOURCE_JNDI_NAME constant
to WildFlyConfigurationExtension, which could initialize it differently
depending on various (static) methods, and could then decide to use the
DataSourceConfigurator or not.