HSEARCH-3246 Skip coverage reporting in legacy tests
This is necessary if we want coveralls to correctly report coverage
changes between master and pull requests.
Without this change, the coverage on master refers to both new and
legacy tests, while the coverage on PRs refers to new (Search 6) tests
only, and those two numbers obviously cannot be compared.