HSEARCH-2554 Test the use of ContainerValueExtractor in @Field
ContainerValueExtractors were added as part of the Search 6 groundwork.

The default behavior (exhibited in
FieldDefaultContainerValueExtractorIT) is to apply the default container
value extractors, and then the value bridge (either the default one or
the explicitly provided one). In short a List<String> will be applied
the IterableElementExtractor, a Map<String, List<String>> will be
applied a MapValueExtractor, then an IterableElementExtractor, and so
on.

But users may also specify container value extractors explicitly (see
FieldExplicitContainerValueExtractorIT).