HSEARCH-3334 Replace usage of deprecated nexus-maven-plugin with nexus-staging-maven-plugin
I could not find any documentation for the older plugin, so I can't tell
what it used to do exactly, just that it was replaced by the
nexus-staging-maven-plugin.

The newer plugin provides similar features to the maven-deploy-plugin,
with two main improvements:

1. It implements deferred deployment, waiting until the whole project
built successfully before it sends the artifacts to the Nexus repository.
This should in most cases prevent "partial" deployments where we deploy
the engine, for example, but then some integration tests fail and we
don't deploy other modules.
2. For non-snapshot versions (where staging is actually relevant), it
will close the staging repository automatically after all artifacts
have been uploaded, or drop it automatically if a failure occurs during
uploads.

Also, we should investigate what the newer plugin provides: it seems we
should be able to orchestrate releases fully from a Jenkins pipeline if
we used the full power of the plugin. That's beyond the scope of this
commit, though.