HSEARCH-3379 Clean up the now unnecessary references to byte-buddy
We don't need to declare the byte-buddy dependency in the OSGi feature
anymore (since we rely on the ORM-provided feature file), so the
byte-buddy version property isn't necessary for OSGi testing anymore.

As to other references to byte-buddy, they are apparently obsolete:

* Hibernate ORM already provides this dependency transitively, so we
don't need to declare it explicitly, especially since we don't use it
directly ourselves
* The "hibernate.bytecode.provider" setting is unnecessary, its default
value is already "bytebuddy".