HSEARCH-3115 Rename objects providing callbacks allowing to configure a part of Hibernate Search to "configurers"
So far we only have three such interfaces: "ElasticsearchAnalysisConfigurer",
"LuceneAnalysisConfigurer" and "HibernateOrmSearchMappingConfigurer".

There are multiple reasons for this renaming.

First, because "contributor" is as generic as it gets, and "definition provider"
as well, so another generic name cannot be worse.

Second, because we should be consistent with how we name interfaces, so
as to avoid perplexing our users. We previously had a "definition
provider" when it comes to analysis, and a "mapping provider" when it
comes to ORM mapping...

Third, because the name for the ORM mapping configurer was a bit wrong:
it was named "HibernateOrmSearchMappingProvider", even though the object does
not provide a mapping, it merely configures the mapping and expects
Hibernate Search to build the mapping.