HSEARCH-3335 Use moduleSets to organize the distribution bundles a bit without requiring to list dependencies explicitly
There are two main advantages of moduleSets over "root" dependencySets:

1. First and foremost, moduleSets don't suffer from the same "implicit
exclusion" as "root" dependencySets. With "root" dependencySets, each
dependencySet implicitly excludes any dependency that was included in
any of the previous dependencySets. Which currently is a nice "feature",
but it cannot be disabled it will cause problems at some point, for
example if we have two backends with a common dependency: the directory
for one of the backends will miss this dependency, and users may end up
with classpath issues.
2. With moduleSets, we can structure the bundle in such a way that our
modules are located in a given directory, while the dependencies are
located in another directory. We don't *have* to, but we can.