HSEARCH-3347 Remove the link to the getting started guide from the README
We will move the getting started guide to first section of the reference
documentation, at least temporarily. This makes it easier to provide the
most appropriate "getting started" guide for each version. And we
already have a "Getting started" section there anyway.

In the future, maybe we could publish a separate getting started guide,
with only the "getting started" section from the reference
documentation. But for now, this will do.