HSEARCH-3296 Use one specific context interface per bridge/converter operation
This allows to de-couple the evolution of the context passed to each
method. This way, if we need, one day, to expose some methods to one
operation, but not to another, we will be able to do so. With a unique
interface for all operations, it wouldn't have been possible.