HSEARCH-3398 Move the implementation of IndexSearchTarget to the engine
... and only expect backends to implement SearchTargetContext.

This makes things easier in the next commits when we make the search
query result definition APIs specific to each mapper.

Besides, this looks like a good idea in any case, since it makes the
backend code simpler and allows us to remove SPIs because they are no
longer relevant.