HSEARCH-3407 Correctly fetch the value of multi-valued boolean fields in the Elasticsearch integration
We used to rely on TwoWayFieldBridge, but this only works when there is
a single value for the field, because we just tell the bridge "fetch the
value for the field with this name", and the bridge doesn't know which
value to fetch, so it'll just return the first one.

We solve this by fetching the raw (string) value ourselves, and passing
the value to the bridge directly.

This means custom bridges not implementing TwoWayStringBridge will not
work. But I don't see the point of custom bridges in this case anyway,
since we only detect the boolean type when the property itself is of
type boolean or Boolean, so users may as well rely on the default
bridge.