HSEARCH-3386 Move the responsibility of checking bean instance types to the BeanReference
Doing that, we simplify the BeanReference configuration properties and
the bean resolving in general.

Note that this required to give up on @Factory annotations.
The type passed to Hibernate Search when referencing a bridge, a filter,
etc. must now always implement the requested interface, we can't just
pass a factory class that doesn't implement anything.
It's not a great loss, though: see HSEARCH-3055 for more information.