HSEARCH-3170 Pass a build context to marker builders
First because it seems to make sense, considering these builders are
very similar to bridge builders, and we may want to pass more context in
the future.

Second, because we will need access to the bean provider in order to
implement the marker builder lifecycle (releasing in particular) properly
in the next commits.