HSEARCH-3170 Delay instantiation of bridge/marker builders until when they are actually needed
This makes resource management easier: we will get the builder beans
each time we need them, thus will be able to release them each time
we use them, too.

We may want something more optimized in the future, but for now this
will do just fine. Especially because it's just bootstrap, and bridge
builders are not expected to be very heavy most of the time.