HSEARCH-3170 Enclose beans retrieved from the provider in BeanHolders to allow early release
This way, we don't need to hold a list of all beans and release them at
shutdown. This is especially interesting for bootstrap-only beans, such
as configurers, bridge builders, and so on, which we used to keep
alive during the whole time Hibernate Search was running, even though
they were only needed during bootstrap.